Refactor project to have better Separation of Concerns and project structure.
