+++ src/gsc/Models/Json/SystemInfoDTOJsonContext.cs
@@ -1,0 +1,7 @@
+
+using System.Text.Json.Serialization;
+
+[JsonSerializable(typeof(SystemInfoDTO))]
+internal partial class SystemInfoDTOJsonContext : JsonSerializerContext
+{
+}
