+++ src/gsc/Models/MemoryStateDto.cs
@@ -1,0 +1,6 @@
+    public record MemoryStateDto
+    {
+        public float Used { get; set; }
+        public float Available { get; set; }
+        public float Load { get; set; }
+    }
