+++ src/gsc/Models/PsuInfoDto.cs
@@ -1,0 +1,6 @@
+public class PsuInfoDto
+{
+    public string Name { get; set; }
+    public Dictionary<string, float> Power { get; set; } = new Dictionary<string, float>();
+    public Dictionary<string, float> Voltage { get; set; } = new Dictionary<string, float>();
+}
